# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CBSE Study Guide API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all CBSE subjects
 */
export const ListSubjectsQueryParams = zod.object({
  grade: zod.coerce.number().optional(),
});

export const ListSubjectsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  grade: zod.number(),
  description: zod.string(),
  icon: zod.string(),
  chapterCount: zod.number(),
  color: zod.string(),
});
export const ListSubjectsResponse = zod.array(ListSubjectsResponseItem);

/**
 * @summary List chapters for a subject
 */
export const ListChaptersParams = zod.object({
  subjectId: zod.coerce.number(),
});

export const ListChaptersResponseItem = zod.object({
  id: zod.number(),
  subjectId: zod.number(),
  title: zod.string(),
  chapterNumber: zod.number(),
  summary: zod.string(),
  lessonStudy: zod.string().nullish(),
  conceptsJson: zod.string().nullish(),
  questionCount: zod.number(),
});
export const ListChaptersResponse = zod.array(ListChaptersResponseItem);

/**
 * @summary Get chapter details with key concepts
 */
export const GetChapterParams = zod.object({
  chapterId: zod.coerce.number(),
});

export const GetChapterResponse = zod.object({
  id: zod.number(),
  subjectId: zod.number(),
  title: zod.string(),
  chapterNumber: zod.number(),
  summary: zod.string(),
  lessonStudy: zod.string().nullish(),
  conceptsJson: zod.string().nullish(),
  keyConcepts: zod.array(
    zod.object({
      id: zod.number(),
      title: zod.string(),
      explanation: zod.string(),
    }),
  ),
  questionCount: zod.number(),
  visuals: zod
    .array(
      zod.object({
        type: zod.enum(["geogebra", "manim"]),
        id: zod.string(),
        title: zod.string(),
        thumbnail: zod.string().optional(),
      }),
    )
    .optional(),
  videos: zod.array(
    zod.object({
      id: zod.number(),
      chapterId: zod.number(),
      youtubeVideoId: zod.string(),
      title: zod.string(),
      description: zod.string().nullish(),
      displayOrder: zod.number(),
    }),
  ),
});

/**
 * @summary List videos for a chapter
 */
export const ListChapterVideosParams = zod.object({
  chapterId: zod.coerce.number(),
});

export const ListChapterVideosResponseItem = zod.object({
  id: zod.number(),
  chapterId: zod.number(),
  youtubeVideoId: zod.string(),
  title: zod.string(),
  description: zod.string().nullish(),
  displayOrder: zod.number(),
});
export const ListChapterVideosResponse = zod.array(ListChapterVideosResponseItem);

/**
 * @summary List practice questions for a chapter
 */
export const ListPracticeQuestionsParams = zod.object({
  chapterId: zod.coerce.number(),
});

export const ListPracticeQuestionsQueryParams = zod.object({
  difficulty: zod.enum(["easy", "medium", "hard"]).optional(),
});

export const ListPracticeQuestionsResponseItem = zod.object({
  id: zod.number(),
  chapterId: zod.number(),
  question: zod.string(),
  difficulty: zod.string(),
  type: zod.string(),
  options: zod.array(zod.string()).nullish(),
  answer: zod.string(),
  explanation: zod.string(),
  hint: zod.string().nullish(),
  stepByStepSolution: zod.string().nullish(),
});
export const ListPracticeQuestionsResponse = zod.array(
  ListPracticeQuestionsResponseItem,
);

/**
 * @summary Get dashboard summary with subject counts and recent activity
 */
export const GetDashboardResponse = zod.object({
  totalSubjects: zod.number(),
  totalChapters: zod.number(),
  totalQuestions: zod.number(),
  gradeSummaries: zod.array(
    zod.object({
      grade: zod.number(),
      subjectCount: zod.number(),
      totalChapters: zod.number(),
    }),
  ),
  recentChapters: zod.array(
    zod.object({
      id: zod.number(),
      subjectId: zod.number(),
      title: zod.string(),
      chapterNumber: zod.number(),
      summary: zod.string(),
      lessonStudy: zod.string().nullish(),
      conceptsJson: zod.string().nullish(),
      questionCount: zod.number(),
    }),
  ),
});

/**
 * @summary Get content generation statistics
 */
export const GetContentStatsResponse = zod.object({
  totalChapters: zod.number(),
  chaptersWithLesson: zod.number(),
  chaptersWithQuestions: zod.number(),
  totalQuestions: zod.number(),
  coveragePercent: zod.number().nullish(),
  chaptersWithVideos: zod.number(),
  totalVideos: zod.number(),
});

/**
 * @summary Get completed chapter IDs for a session
 */
export const GetProgressQueryParams = zod.object({
  sessionId: zod.coerce.string(),
});

export const GetProgressResponse = zod.object({
  completedChapterIds: zod.array(zod.number()),
});

/**
 * @summary Mark a chapter as complete
 */
export const MarkChapterCompleteBody = zod.object({
  sessionId: zod.string(),
  chapterId: zod.number(),
});

/**
 * @summary Remove completion mark from a chapter
 */
export const UnmarkChapterCompleteParams = zod.object({
  chapterId: zod.coerce.number(),
});

export const UnmarkChapterCompleteQueryParams = zod.object({
  sessionId: zod.coerce.string(),
});

export const UnmarkChapterCompleteResponse = zod.object({
  completedChapterIds: zod.array(zod.number()),
});

/**
 * @summary List all conversations
 */
export const ListOpenaiConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListOpenaiConversationsResponse = zod.array(
  ListOpenaiConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateOpenaiConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get conversation with messages
 */
export const GetOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetOpenaiConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List messages in a conversation
 */
export const ListOpenaiMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListOpenaiMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListOpenaiMessagesResponse = zod.array(
  ListOpenaiMessagesResponseItem,
);

/**
 * @summary Send a text message and receive a streaming text response
 */
export const SendOpenaiMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendOpenaiMessageBody = zod.object({
  content: zod.string(),
});